Platform and functions
Changed the platform to Android
Created variable in database to save the `lock` state of a specific soil slot.
Get the values from database with a script at the beginning of the game.
If the slot is locked place a Lock icon on the specific slot otherwise open the planting menu.
Created a function for unlocking a slot.
(not finished yet)