Thirsty and dying plants
Added new plants functionalities to become thirsty after some time and the player must water them, also if the player doesnt water the plants in time they will die and the player must remove them using a shovel button.
Edited .php scripts to save data about how thirsty a plant is or if it is dead or not.